Hella neat.
Cleaned up the files, big time.
Changes to the config. Now able to rename and create new trails right in the config.
The name permissions of the trail are based on the name of the trail, so if you rename a trail the permissions have to be changed as well.
Added blackdust trail
added footsteps trail
added blackpotion trail
added automatic pagination to the GUI
Minor improvement to /trail
improved performance (due to there being vastly less config calls)
Renamed "Methodes" to "Methods"
Shifted trail logic to the Trail classes.
Condensed activetrails lists into a single map
Moved version number up to 3.11